splitting out msgSerializer
